Refactoring groupId
